fix: 🐛 cwl bug patches (#55)
* wip: :construction: :bug: patched count on modifying cwl file

* refactor: :recycle: update applycwltemplate fn to make it easier to read

* refactor: :recycle: update on cwl workflow, improved handling and validation

* wip: :construction: bug catching cwl workflow

* feat: :sparkles: final touches for cwl workflow

* feat: :sparkles: boolean db field for fair items existence check

* refactor: :recycle: apply<fairItem>template moved to fair item file
